Create Mergesort.py
Merge Sort in Python:

Merge Sort is a popular sorting algorithm that follows the divide-and-conquer approach. It divides an unsorted list into smaller sublists until each sublist consists of one element, which is already sorted. Then, it merges these sublists back together in a sorted manner. Merge Sort is efficient and stable, making it a reliable choice for sorting large datasets.